database migration/change how I reload measurement data
use the database structure that is changed on commit c2834df7f1a
and the HHSmiddelware commit ba7045dd4a16b

measurement data
change the approach of how i check for updates for the NetworkOverview page.

first approach is get all new values from the database from a user and than go to the appropriate label and change the text.

there was a bug in how the data is pulled from the database.
only pulling the most resent data from all sensors instead of getting the newest values